requirements.txt
Flask==1.1.2
torch==1.4.0
torchvision==0.5.0
opencv-python==4.2.0.34
ffmpeg-python==0.2.0
scikit-image==0.16.2
pillow==7.0.0
scipy==1.2.1
tqdm==4.43.0
munch==2.5.0
requests==2.24.0